backend/requirements.txt
fastapi
uvicorn
python-multipart
sentence-transformers
langchain
langchain-community
langchain-chroma
langchain-huggingface
pypdf
langchain-openai
openai>=1.30.0
python-dotenv>=1.0.1
cryptography
pymysql
sqlalchemy
python-jose[cryptography]
pydantic
python-multipart
pandas
bcrypt==4.1.2
passlib[bcrypt]>=1.7.4
typing-extensions>=4.12.2
email-validator
pydantic[email]
